Store the payload in a dedicated section
To my knowledge the Zig language does not support variable attributes
that could be used to determine the section in which data is stored.
Therefore, the payload data is defined in a C source file that assigns
the required attributes to store the data in a dedicaed payload section.
The custom linker script will ensure that this section is stored at the
end of the binary.